refactor: split runtime support modules
